Add files via upload
How to run: python You can run the file from file explorer and also add it to your vscode. Current bugs: Cant back out of the screen when in the cost estimation/calendar menu. Service history page has not been completed. Overall this is just the beginning and a lot of visual changes will be made.